kill buttons ts file, engagement button working, eliminate separate style props
