+++ client/scripts/views/components/component_kit/cw_button.tsx
@@ -6,24 +6,27 @@ import 'components/component_kit/cw_button.scss';
 import { ButtonType, ComponentType, State, StyleProps } from './types';
 import { getClasses } from './helpers';
 
-type ButtonStyleProps = {
+export type ButtonStyleProps = {
   buttonType: ButtonType;
 } & StyleProps;
 
 export type ButtonProps = {
-  styleProps: ButtonStyleProps;
   label: string;
   onclick: (e?: MouseEvent) => void;
-};
+} & ButtonStyleProps;
 
 export const CWButton: m.Component<ButtonProps, State> = {
   view: (vnode) => {
-    const { onclick, label, styleProps } = vnode.attrs;
+    const { onclick, label, disabled, className, buttonType } = vnode.attrs;
     return (
       <button
-        class={getClasses(ComponentType.Button, styleProps)}
+        class={getClasses(ComponentType.Button, {
+          disabled,
+          className,
+          buttonType,
+        })}
         onclick={onclick}
-        disabled={styleProps.disabled}
+        disabled={disabled}
       >
         <span>{label}</span>
       </button>
