+++ client/scripts/views/components/component_kit/cw_gradient_button.tsx
@@ -1,19 +1,24 @@
 /* @jsx m */
 
 import m from 'mithril';
-import 'components/component_kit/gradient_button.scss';
+import 'components/component_kit/cw_gradient_button.scss';
 
-import { ComponentType, State, ButtonProps } from './types';
+import { ComponentType, State } from './types';
 import { getClasses } from './helpers';
+import { ButtonProps } from './cw_button';
 
 export const CWGradientButton: m.Component<ButtonProps, State> = {
   view: (vnode) => {
-    const { onclick, label, styleProps } = vnode.attrs;
+    const { onclick, label, disabled, className, buttonType } = vnode.attrs;
     return (
       <button
-        class={getClasses(ComponentType.GradientButton, styleProps)}
+        class={getClasses(ComponentType.GradientButton, {
+          disabled,
+          className,
+          buttonType,
+        })}
         onclick={onclick}
-        disabled={styleProps.disabled}
+        disabled={disabled}
       >
         <span>{label}</span>
       </button>
