+++ client/styles/components/component_kit/cw_engagement_button.scss
@@ -9,34 +9,43 @@
   border-radius: 20px;
   height: 40px;
   font-weight: 500;
+
   .Icon {
     margin-right: 12px;
   }
-  .label {
+
+  span {
     padding-top: 5.5px;
   }
+
   &:hover {
     color: $dark-purp;
     background: $xlite-purp;
     border: 1px solid $dark-purp;
+
     .Icon path {
       stroke: $dark-purp;
     }
   }
+
   &.sm {
     padding: 12px 20px;
   }
+
   &.lg {
     padding: 12px 44px;
   }
+
   &.disabled {
     color: $disable-gray;
     background: $white;
     cursor: default;
+
     &:hover {
       color: $disable-gray;
       border: 1px solid $disable-gray;
       background: $white;
+
       .Icon path {
         stroke: $disable-gray;
       }
